requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas>=2.2.2  # Version flexible pour compatibilité Windows (wheel précompilé)
python-dotenv==1.0.0
typing-extensions==4.9.0

# Inference
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
